backend/pyproject.toml
[tool.poetry]
name = "ai-code-review-backend"
version = "0.1.0"
description = "Backend for ai-code-review-squad"
authors = ["YOUR_ORG <info@your_org.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.5"
pydantic = "^2.3.0"
redis = "^5.0.1"
celery = "^5.3.4"
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
prometheus-client = "^0.17.1"
opentelemetry-api = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
opentelemetry-instrumentation-fastapi = "^0.39b0"
opentelemetry-instrumentation-sqlalchemy = "^0.39b0"
opentelemetry-instrumentation-celery = "^0.39b0"
python-multipart = "^0.0.6"
requests = "^2.31.0"
pandas = "^2.0.3"
pyarrow = "^12.0.1"
great-expectations = "^0.17.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
ruff = "^0.0.289"
black = "^23.3.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"